fix: Fix render error (#123)
